backups/2025-12-05_18-42-58/files/scripts/orchestrator/requirements.txt
# Orchestrator Requirements

# GitHub Integration
PyGithub>=2.1.1

# JSON Schema Validation (optional, for advanced validation)
jsonschema>=4.19.0

# Environment Variables
python-dotenv>=1.0.0

# YAML Parsing (if needed for configs)
PyYAML>=6.0

# Date/Time utilities (if needed)
python-dateutil>=2.8.2

